Removed browserify
Will be missed, but the CommonJS pattern is a forced fit with both the
Rails asset pipeline and AngularJS.  AngularJS defeats the purpose of
much of the benefit of the pattern given the nature of injection and
module lookup, and the fact that it requires shimming both itself and
jQuery regardless.  It leaves little excuse for the extra moving parts
and the inconvenience of running a filesystem watch to keep the bundle
up to date given that the asset pipeline would otherwise handle bundle
compilation on request.