Remove min crop size
This was in place to protect the mill from uploading images that would
result in excessively low-res crops, but it turns out that this causes
a UI break and a resulting exception on the server if the uploaded image
is smaller than the min crop size.  So, the min crop size is removed for
now.

The more appropriate solution would be to disallow images below certain
dimensions from even being uploaded, but that can be for another time.