added "duplicate" func for copying existing cways
The new action for color ways now accepts a "duplicate_id" param that
designates an existing color way to inherit from.  Complicated slightly
by the image setup, the new form will now show the duplicate's image for
cropping purposes if it exists, and correspondingly the image will be
copied over in the create action.  If a new image is uploaded it will
override this behavior.